# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Cupreous Dating App"
    },
    "host": "http://54.152.239.30/:9118",
    "basePath": "/",
    "paths": {
...
        "/user/updateuser/{id}": {
            "put": {
                "tags": ["user"],
                "summary": "Update user",
                "description": "API for updating user information only for logged in Users.",
                "operationId": "updateUser",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "name that need to be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Updated user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/user/forgot": {
            "post": {
                "tags": ["user"],
                "summary": "This is to Recover Forgotten Password",
                "description": "API for getting forgotten Password.",
                "operationId": "forgotPassword",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "email",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ForgotPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/user/password/reset/{token}": {
            "put": {
                "tags": ["user"],
                "summary": "Reset user Password",
                "description": "API for updating user changing password only for logged in Users.",
                "operationId": "updateUser",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Updated user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ResetPassword"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ResetPassword_Response"
                        }
                    }
                }
            }
        },
...
    },
...
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "contactNo": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                
                "gender": {
                    "type": "string"
                },
                "height": {
                    "type": "string"
                },
               
                "maritalStatus": {
                    "type": "string"
                },
                "qualification": {
                    "type": "string"
                },
                
                "password": {
                    "type": "String"
                },
                "confirmPassword": {
                    "type": "string"
                },
                
                "ethnicity": {
                    "type": "string"
                }
               
            }
        },
...
        "ForgotPassword": {
            "type": "object",
            "properties": {
                "email": {
                    "type":"string"
                }
            }
        },
...
        "ResetPassword": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "required": true
                },
                "confirmPassword": {
                    "type": "string",
                    "required": true
                }
            }
        },
        "ResetPassword_Response": {
            "type": "object",
            "properties": {
                "User": {
                    "$ref": "#/definitions/User"
                },
                "Success": true,
                "token": "String"
            }
        },
...
    },
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"]
}